* ANDROID BUGFIX: Trying to fix segfault issue with Android. Possibly due to logging inputs.

+++ include/madara/transport/udp/UdpRegistryClient.cpp
@@ -218,8 +218,8 @@ madara::transport::UdpRegistryClient::send_register (void)
     {
       madara_logger_log (context_.get_logger (), logger::LOG_MAJOR,
         "%s:" \
-        " Sending register of size %ld to %s\n",
-        print_prefix, result, i->first.c_str ());
+        " Sending register of size %d to %s\n",
+        print_prefix, (int)result, i->first.c_str ());
 
       int send_attempts = -1;
       ssize_t actual_sent = -1;
@@ -241,7 +241,7 @@ madara::transport::UdpRegistryClient::send_register (void)
 
           madara_logger_log (context_.get_logger (), logger::LOG_MAJOR,
             "%s:" \
-            " Sent register of size %" PRIu64 " to %s\n",
+            " Sent register of size %d to %s\n",
             print_prefix, (int)actual_sent, i->first.c_str ());
 
           send_monitor_.add ((uint32_t)actual_sent);
@@ -385,7 +385,7 @@ madara::transport::UdpRegistryClient::send_data (
 
                 madara_logger_log (context_.get_logger (), logger::LOG_MAJOR,
                   "%s: Send result was %d of %d byte fragment to %s\n",
-                  print_prefix, actual_sent, frag_size, addr->first.c_str ());
+                  print_prefix, (int)actual_sent, (int)frag_size, addr->first.c_str ());
 
                 if (actual_sent > 0)
                 {
@@ -473,8 +473,8 @@ madara::transport::UdpRegistryClient::send_data (
           {
             madara_logger_log (context_.get_logger (), logger::LOG_MAJOR,
               "%s:" \
-              " Sending packet of size %ld to %s\n",
-              print_prefix, result, i->first.c_str ());
+              " Sending packet of size %d to %s\n",
+              print_prefix, (int)result, i->first.c_str ());
 
             int send_attempts = -1;
             ssize_t actual_sent = -1;
@@ -496,7 +496,7 @@ madara::transport::UdpRegistryClient::send_data (
 
                 madara_logger_log (context_.get_logger (), logger::LOG_MAJOR,
                   "%s:" \
-                  " Sent packet of size %" PRIu64 " to %s\n",
+                  " Sent packet of size %d to %s\n",
                   print_prefix, (int)actual_sent, i->first.c_str ());
 
                 send_monitor_.add ((uint32_t)actual_sent);
